+++ apps/sample-next/src/components/home/device-registration-card.tsx
@@ -3,6 +3,8 @@ import { getCurrentDevice, registerDevice, RegisterDeviceOptions } from "notific
 import { Button } from "@/components/button";
 import { Card, CardActions, CardContent, CardHeader } from "@/components/card";
 import { InputField } from "@/components/input-field";
+import { useOnDeviceRegistered } from "@/notificare/hooks/events/core/device-registered";
+import { logger } from "@/utils/logger";
 
 export function DeviceRegistrationCard() {
   const device = getCurrentDevice();
@@ -10,6 +12,11 @@ export function DeviceRegistrationCard() {
   const [userId, setUserId] = useState<string>(device?.userId ?? "");
   const [userName, setUserName] = useState<string>(device?.userName ?? "");
   const [loading, setLoading] = useState<boolean>(false);
+
+  useOnDeviceRegistered((device) => {
+    setUserId(device.userId ?? "");
+    setUserName(device.userName ?? "");
+  });
 
   const onRegisterClick = useCallback(() => {
     setLoading(true);
@@ -22,7 +29,7 @@ export function DeviceRegistrationCard() {
     registerDevice(options)
       .then(() => setLoading(false))
       .catch((e) => {
-        console.error(`Unable to register device: ${e}`);
+        logger.error(`Unable to register device: ${e}`);
         setLoading(false);
       });
   }, [userId, userName]);