+++ packages/notificare-push/src/internal/push-component.ts
@@ -27,6 +27,8 @@ import {
   setRemoteNotificationsEnabled,
   storeAllowedUI,
 } from './storage/local-storage';
+import { hideFloatingButton } from './ui/floating-button';
+import { hideOnboarding } from './ui/onboarding';
 
 /* eslint-disable class-methods-use-this */
 export class PushComponent extends Component {
@@ -74,7 +76,9 @@ export class PushComponent extends Component {
   }
 
   async launch(): Promise<void> {
-    if (getRemoteNotificationsEnabled() && getPushPermissionStatus() === 'granted') {
+    // An undefined getRemoteNotificationsEnabled() likely means the user or the app
+    // tampered with the local storage. We can recover based on the browser permission.
+    if (getRemoteNotificationsEnabled() !== false && getPushPermissionStatus() === 'granted') {
       try {
         logger.debug('Automatically enabling remote notifications.');
         await enableRemoteNotifications();
@@ -87,6 +91,8 @@ export class PushComponent extends Component {
   }
 
   async unlaunch(): Promise<void> {
+    this.removeOnboardingElements();
+
     localStorage.removeItem('re.notifica.push.first_registration');
     localStorage.removeItem('re.notifica.push.onboarding_last_attempt');
 
@@ -99,7 +105,7 @@ export class PushComponent extends Component {
       await disableWebPushNotifications();
     }
 
-    setRemoteNotificationsEnabled(false);
+    setRemoteNotificationsEnabled(undefined);
     storeAllowedUI(undefined);
   }
 
@@ -215,5 +221,10 @@ export class PushComponent extends Component {
 
     const notification = await fetchNotification(notificationId);
     notifyNotificationOpened(notification);
+  }
+
+  private removeOnboardingElements() {
+    hideOnboarding();
+    hideFloatingButton();
   }
 }